Moved helpers.js to better place, cleaned up code
